lab13

+++ web/app/views.py
@@ -14,7 +14,7 @@ from app import login_manager
 
 from app.models.contact import Contact
 from app.models.blogentry import BlogEntry
-from app.models.authuser import AuthUser, PrivateContact
+from app.models.authuser import AuthUser, PrivateContact, PrivateBlog
 import datetime
 
 
@@ -95,17 +95,18 @@ def lab10_phonebook():
     return render_template('lab12/lab10_phonebook.html')
 
 
-
 @app.route("/lab10/contacts")
 @login_required
 def lab10_db_contacts():
     # db_contacts = Contact.query.all()
-    db_contacts = PrivateContact.query.filter(PrivateContact.owner_id == current_user.id)
+    # db_contacts = PrivateContact.query.filter(PrivateContact.owner_id == current_user.id)
+    db_contacts = PrivateContact.query.all()
     contacts = list(map(lambda x: x.to_dict(), db_contacts))
     app.logger.debug("DB Contacts: " + str(contacts))
 
 
     return jsonify(contacts)
+
 
 @app.route('/lab10/remove_contact', methods=('GET', 'POST'))
 @login_required
@@ -127,6 +128,7 @@ def lab10_remove_contacts():
 
 
 @app.route('/lab11', methods=('GET', 'POST'))
+@login_required
 def lab11_bootstrap():
     if request.method == 'POST':
         result = request.form.to_dict()
@@ -151,19 +153,20 @@ def lab11_bootstrap():
                 break
             validated_dict[key] = value
         if validated:
-            # utc_dt = datetime.datetime.now()
-            # utc_dt_1 = utc_dt.strftime("%c")
-            # validated_dict['date'] = utc_dt_1
+            
+            
             
             app.logger.debug('validated dict: ' + str(validated_dict))
             # if there is no id: create a new contact entry
             if not id_:
-                entry = BlogEntry(**validated_dict)
+                validated_dict['owner_id'] = current_user.id
+                entry = PrivateBlog(**validated_dict)
                 app.logger.debug(str(entry))
                 db.session.add(entry)
             # if there is an id already: update the contact entry
             else:
-                blogentry = BlogEntry.query.get(id_)
+                validated_dict['owner_id'] = current_user.id
+                blogentry = PrivateBlog.query.get(id_)
                 blogentry.update(**validated_dict)
 
 
@@ -171,32 +174,43 @@ def lab11_bootstrap():
 
 
         return lab11_db_blog()  
-    return app.send_static_file('lab11_microblog.html')
+    return render_template('lab12/lab11_microblog.html')
 
 @app.route("/lab11/blog")
 def lab11_db_blog():
     blogEntrys = []
     # https://stackoverflow.com/questions/15791760/how-can-i-do-multiple-order-by-in-flask-sqlalchemy
-    db_blogentry = BlogEntry.query.order_by(BlogEntry.date_update.desc()).all()
-    
+    db_blogentry = PrivateBlog.query.order_by(BlogEntry.date_update.desc()).all()
     blogEntrys = list(map(lambda x: x.to_dict(), db_blogentry))
     app.logger.debug("DB blog: " + str(blogEntrys))
     
     return jsonify(blogEntrys)
 
+
+@app.route("/lab11/user")
+def lab11_db_user():
+    blogEntrys = []
+    db_blogentry = AuthUser.query.all()
+    blogEntrys = list(map(lambda x: x.to_dict(), db_blogentry))
+    return jsonify(blogEntrys)
+
+
 @app.route('/lab11/remove_blog', methods=('GET', 'POST'))
+@login_required
 def lab11_remove_blog():
     app.logger.debug("LAB11 - REMOVE")
     if request.method == 'POST':
         result = request.form.to_dict()
         id_ = result.get('id', '')
-        try:
-            blog = BlogEntry.query.get(id_)
-            db.session.delete(blog)
-            db.session.commit()
-        except Exception as ex:
-            app.logger.debug(ex)
-            raise
+        contact = PrivateBlog.query.get(id_)
+        if contact.owner_id == current_user.id:
+            try:
+                blog = PrivateBlog.query.get(id_)
+                db.session.delete(blog)
+                db.session.commit()
+            except Exception as ex:
+                app.logger.debug(ex)
+                raise
     return lab11_db_blog()
 
 
@@ -245,8 +259,6 @@ def lab12_login():
 
 @app.route('/lab12/signup', methods=('GET', 'POST'))
 def lab12_signup():
-
-
     if request.method == 'POST':
         result = request.form.to_dict()
         app.logger.debug(str(result))
@@ -255,6 +267,7 @@ def lab12_signup():
         validated_dict = {}
         valid_keys = ['email', 'name', 'password']
 
+        
 
         # validate the input
         for key in result:
@@ -305,7 +318,6 @@ def lab12_signup():
 
 
 
-
 def gen_avatar_url(email, name):
     bgcolor = generate_password_hash(email, method='sha256')[-6:]
     color = hex(int('0xffffff', 0) -
@@ -330,4 +342,74 @@ def lab12_logout():
     logout_user()
     return redirect(url_for('lab12_index'))
 
+@app.route('/lab12/edit', methods=('GET', 'POST'))
+@login_required
+def lab12_edit():
+    
+    app.logger.debug('validated dict: ' + str(AuthUser.query.get(current_user.id)))
+    app.logger.debug(AuthUser.query.all())
+    app.logger.debug(PrivateBlog.query.all())
+    db_user = AuthUser.query.all()
+    user = list(map(lambda x: x.to_dict(), db_user))
+    app.logger.debug(user)
+    db_contacts = PrivateBlog.query.all()
+    contacts = list(map(lambda x: x.to_dict(), db_contacts))
+    app.logger.debug(contacts)
+    if request.method == 'POST':
+        result = request.form.to_dict()
+        app.logger.debug(str(result))
+ 
+        validated = True
+        validated_dict = {}
+        valid_keys = ['email', 'name', 'password']
 
+        
+
+        # validate the input
+        for key in result:
+            app.logger.debug(str(key)+": " + str(result[key]))
+            # screen of unrelated inputs
+            if key not in valid_keys:
+                continue
+
+
+            value = result[key].strip()
+            if not value or value == 'undefined':
+                validated = False
+                break
+            validated_dict[key] = value
+
+        if validated:
+            app.logger.debug('validated dict: ' + str(validated_dict))
+            email = validated_dict['email']
+            name = validated_dict['name']
+            password = validated_dict['password']
+
+            auth_users = AuthUser.query.get(current_user.id)
+            validated_dict['password'] = generate_password_hash(password, method='sha256')
+            avatar_url = gen_avatar_url(email, name)
+            validated_dict['avatar_url'] = avatar_url
+            # add the new user to the database
+            auth_users.update(**validated_dict)
+            db.session.commit()
+            app.logger.debug('validated dict: ' + str(AuthUser.query.get(current_user.id)))
+        return redirect(url_for('lab12_profile'))  
+    return render_template('lab12/edit.html')
+
+
+@app.route('/lab12/bedit', methods=('GET', 'POST'))
+@login_required
+def lab12_bedit():
+    if request.method == 'POST':
+
+        password = request.form.get('password')
+        
+        if not check_password_hash(current_user.password, password):
+            flash('Please check your password details and try again.')
+
+            return render_template('lab12/bedit.html')
+
+
+       
+        return redirect(url_for('lab12_edit'))
+    return render_template('lab12/bedit.html')
